+++ service/src/utils/mod.rs
@@ -2,5 +2,6 @@ pub mod github;
 pub mod wget;
 pub mod download;
 pub mod zip_extract;
+pub mod fs;